+++ oap-server/server-starter-boot/server-starter-boot-core/src/main/java/com/piesat/skywalking/vo/NetworkVo.java
@@ -3,12 +3,14 @@ package com.piesat.skywalking.vo;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class NetworkVo {
     @ApiModelProperty(value = "时间")
     private String timestamp;
     @ApiModelProperty(value = "流量 输入速度 kb/s")
-    private double inSpeed;
+    private BigDecimal inSpeed;
     @ApiModelProperty(value = "流量 输出速度 kb/s")
-    private double outSpeed;
+    private BigDecimal outSpeed;
 }
